"make dist" fails when workstation has no udev package
Unable to create a tar file for the vmmouse package or run distcheck for that
matter. In tools, when the dev package is not installed, the makefile code
under HAS_UDEV_RULES_DIR is commented out. That leaves $(udev_DATA) empty
which causes EXTRA_DIST to be missing a file.

The solution is to spell out the file name. The name can never change as the
content of the tarball must be the same for everyone creating a tar file from
git, regardless of their workstation configuration.

Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>